split coverage graph into its own dir
